Moving functions into util files
