+++ app.py
@@ -11,12 +11,9 @@ import dash_bootstrap_components as dbc
 import base64
 import io
 import pandas as pd
-import plotly
-import plotly.graph_objects as go
 from dash.dependencies import Input, Output, State
-import math
-import numpy as np
-import plotly.express as px
+from uiutils import update_dropdowns, serve_layout
+from plotutils import generate_plot
 
 external_stylesheets = [dbc.themes.BOOTSTRAP]
 app = dash.Dash(__name__, external_stylesheets=external_stylesheets)
@@ -29,118 +26,9 @@ styles = {
         'overflowX': 'scroll'
     }
 }
-
-# Color dictionary - Mary
-#Dicitonary for colors of pathways
-color_dict={'Integrin':'limegreen',
-            'Blood_Coagulation':'firebrick',
-            'Cytoskeleton':'orangered',
-            'Chemokine_Cytokine_Signaling':'tomato',
-            'Chemokine_Cytokine_Signaling, Cytoskeleton':'royalblue',
-            'Chemokine_Cytokine_Signaling, Cytoskeleton, Integrin':'seagreen',
-            'Chemokine_Cytokine_Signaling, Cytoskeleton, Huntington':'wheat',
-            'Chemokine_Cytokine_Signaling, Huntington, Integrin':'yellowgreen',
-            'Chemokine_Cytokine_Signaling, Cytoskeleton, Huntington, Integrin':'violet',
-            'Huntington, Integrin':'crimson',
-            'Parkinson':'lightseagreen',
-            'Cytoskeleton, Huntington':'aqua',
-            'Chemokine_Cytokine_Signaling, Integrin':'palegreen',
-            'Glycolysis, Huntington':'chocolate',
-            'ATP_Synthesis':'brown',
-            'ATP_Synthesis, Huntington':'burlywood',
-            'Glycolysis':'mediumvioletred',
-            'Glycolysis, Pyruvate_Metabolism':'cadetblue',
-            'Huntington':'goldenrod',
-            'Pyruvate_Metabolism':'darkkhaki',
-            'Pyruvate_Metabolism, TCA_Cycle':'sienna',
-            'TCA_Cycle':'darkred',
-            'De_Novo_Purine_Biosynthesis':'mediumpurple',
-            'None':'slategray'}
-
 # Declare a global dataframe to hold the uploaded information
 df = pd.DataFrame()
 files = {}
-
-def serve_layout():
-    layout = html.Div([
-        # The following dbc.Row contains the 2 column layout where settings are on the left
-        # and the plot is the column to the right
-
-        dbc.Row([
-            dbc.Col([
-                html.Div(className='six columns', children=[
-                    dcc.Upload(
-                        id='upload-data',
-                        children=html.Div([
-                            html.Button('Click Here to Select File(s)')
-                        ]),
-                        style={
-                            'width': '100%',
-                            'height': '60px',
-                            'lineHeight': '60px',
-                            'textAlign': 'center',
-                            'margin': '10px'
-                        },
-                        # Allow multiple files to be uploaded
-                        multiple=True
-                    ),
-                    html.Div(id='file-selector'),
-                    html.Div(id='file-name'),
-                    html.Div(id='output-data-upload'),
-
-                    html.Div(id='dropdown-items'),
-                    html.Label('Plot Axis Scale'),
-                    html.Div(children=[
-                        dcc.RadioItems(
-                            id='scale-radio',
-                            options=[
-                                {'label': 'Linear', 'value': 'lin'},
-                                {'label': 'Log10', 'value': 'log'}
-                            ],
-                            value='lin',
-                            labelStyle={'display': 'block'}
-                        )]),
-                ])
-            ], width={"size": 2}
-
-            ),
-            # Figure plot
-            dbc.Col(
-                html.Div(className='six columns', children=[
-                    dcc.Graph(
-                        id='basic-interactions',
-                        clear_on_unhover=True,
-                        config={
-                            'toImageButtonOptions': {
-                                'format': 'svg',  # one of png, svg, jpeg, webp
-                                'filename': 'custom_image',
-                                'height': 800,
-                                'width': 800,
-                                'scale': 1  # Multiply title/legend/axis/canvas sizes by this factor
-                            },
-
-                        },
-
-                    )
-                ])
-            )
-            # Second row is for the hover data which only has a single, centered column
-        ]),
-        dbc.Row([
-            dbc.Col([
-                html.Div([
-                    dcc.Markdown("""
-                                    **Hover Data**
-                            
-                                    Mouse over values in the graph.
-                                """),
-                    html.Pre(id='hover-data')
-                ])
-            ], width={"size": 6, "offset": 3})
-        ])
-    ])
-    return layout
-
 
 app.layout = serve_layout
 
@@ -204,51 +92,7 @@ def drop_down_updates(file_name):
     global df
     if file_name is not None:
         df = pd.DataFrame.from_dict(files[file_name])
-
-    return [html.Label(f"File: {file_name}"), (
-        html.Label('Row names'),
-        dcc.Dropdown(
-            id='name-dropdown',
-            options=[{'label': name, 'value': name} for name in df.columns],
-            clearable=False,
-            searchable=True
-        ),
-        html.Label('Color by:'),
-        dcc.Dropdown(
-            id='colour-dropdown',
-            options=[{'label': name, 'value': name} for name in df.columns],
-            clearable=False,
-            searchable=True
-        ),
-        html.Label('Positive X'),
-        dcc.Dropdown(
-            id="xpos-dropdown",
-            options=[{'label': name, 'value': name} for name in df.columns],
-            clearable=False,
-            searchable=True,
-        ),
-        html.Label('Positive Y'),
-        dcc.Dropdown(
-            id="ypos-dropdown",
-            options=[{'label': name, 'value': name} for name in df.columns],
-            clearable=False,
-            searchable=True,
-        ),
-        html.Label('Negative X'),
-        dcc.Dropdown(
-            id="xneg-dropdown",
-            options=[{'label': name, 'value': name} for name in df.columns],
-            clearable=False,
-            searchable=True,
-        ),
-        html.Label('Negative Y'),
-        dcc.Dropdown(
-            id="yneg-dropdown",
-            options=[{'label': name, 'value': name} for name in df.columns],
-            clearable=False,
-            searchable=True,
-        ))
-            ]
+    return update_dropdowns(file_name, df)
 
 
 # After selecting columns to plot, create those traces
@@ -264,95 +108,7 @@ def drop_down_updates(file_name):
      State('basic-interactions', 'figure')])
 def create_figure(xpos, ypos, xneg, yneg, scale, name, colour_by, state):
     global df
-    fig = go.Figure()
-    fig.update_layout(height=600, width=600, autosize=False)
-    if None not in [xpos, ypos, xneg, yneg]:
-        # Create a figure for each quadrant
-        fig1 = px.scatter(df, x=xpos, y=ypos, color=colour_by, custom_data=[name],
-                          hover_data={
-                              xpos:False,
-                              ypos:False,
-                              colour_by: False,
-                              name: True
-                            })
-        fig2 = px.scatter(df, x=xneg, y=ypos, color=colour_by, custom_data=[name],
-                          hover_data={
-                              xpos: False,
-                              ypos: False,
-                              colour_by: False,
-                              name: True
-                          })
-        fig2.update_layout(showlegend=False)
-        fig3 = px.scatter(df, x=xneg, y=yneg, color=colour_by, custom_data=[name],
-                          hover_data={
-                              xpos: False,
-                              ypos: False,
-                              colour_by: False,
-                              name: True
-                          })
-        fig3.update_layout(showlegend=False)
-        fig4 = px.scatter(df, x=xpos, y=yneg, color=colour_by, custom_data=[name],
-                          hover_data={
-                              xpos: False,
-                              ypos: False,
-                              colour_by: False,
-                              name: True
-                          })
-        fig4.update_layout(showlegend=False)
-
-        # Set negative columns to negative values
-        for sc in fig2['data']:
-            sc['x'] = sc['x'] * -1
-        for sc in fig3['data']:
-            sc['x'] = sc['x'] * -1
-            sc['y'] = sc['y'] * -1
-        for sc in fig4['data']:
-            sc['y'] = sc['y'] * -1
-        # Find the maximum value from any column and round to neared 100
-        max_val = max(np.concatenate([np.concatenate([abs(sc['x']) for sc in fig1['data']]),
-                                      np.concatenate([abs(sc['y']) for sc in fig1['data']]),
-                                      np.concatenate([abs(sc['x']) for sc in fig2['data']]),
-                                      np.concatenate([abs(sc['y']) for sc in fig2['data']]),
-                                      np.concatenate([abs(sc['x']) for sc in fig3['data']]),
-                                      np.concatenate([abs(sc['y']) for sc in fig3['data']]),
-                                      np.concatenate([abs(sc['x']) for sc in fig4['data']]),
-                                      np.concatenate([abs(sc['y']) for sc in fig4['data']])]))
-        axis_max = int(round(float(max_val) / 100) * 100)
-
-        # Create the final figure with new axis limits
-        fig = go.Figure(data=fig1.data + fig2.data + fig3.data + fig4.data,
-                         layout_xaxis_range=[axis_max * -1, axis_max],
-                         layout_yaxis_range=[axis_max * -1, axis_max])
-
-        # Update the axis ticks so they don't show negative values
-        fig.update_layout(
-            xaxis=dict(
-                tickmode='array',
-                tickvals=np.concatenate([list(range(axis_max * -1, 1, 100)), list(range(0, axis_max + 1, 100))]),
-                ticktext=np.concatenate(
-                    [[x * -1 for x in list(range(axis_max * -1, 1, 100))], list(range(0, axis_max + 1, 100))])
-            ),
-            yaxis=dict(
-                tickmode='array',
-                tickvals=np.concatenate([list(range(axis_max * -1, 1, 100)), list(range(0, axis_max + 1, 100))]),
-                ticktext=np.concatenate(
-                    [[x * -1 for x in list(range(axis_max * -1, 1, 100))], list(range(0, axis_max + 1, 100))])
-            ),
-            width=600,
-            height=600, showlegend=False
-        )
-
-    # Add in axis labels
-
-    fig.add_hline(y=0)
-    fig.add_vline(x=0)
-    fig.update_layout(
-
-                      xaxis_title=f"\u2190{xneg}-----{xpos}\u2192",
-                      yaxis_title=f"\u2190{yneg}-----{ypos}\u2192",
-                      paper_bgcolor='rgba(0,0,0,0)',
-                      plot_bgcolor='rgba(0,0,0,0)')
-    return fig
+    return generate_plot(df, xpos, ypos, xneg, yneg, scale, name, colour_by)
