+++ requirements.txt
@@ -6,10 +6,11 @@ dash-core-components==1.15.0
 dash-html-components==1.1.2
 dash-renderer==1.9.0
 dash-table==4.11.2
-Flask-Compress==1.5.0
+Flask-Compress==1.9.0
 gunicorn==20.0.4
 pandas==1.2.1
 PyYAML==5.4.1
 retrying==1.3.3
 plotly==4.14.3
 dash-bootstrap-components==0.11.2
+numpy==1.20.1
